lib/client: make createConnection callback-driven
Before this commit, the createConnection option to uhura.Client was
expected to return a net.Socket-like object.  This commit makes it
take a callback that can be invoked synchronously or asynchronously.

Makes it possible to (for example) make a HTTP CONNECT request to a
proxy first before passing on the connection to Uhura.