lib/client: conditionally unref reconnect timer
Commit ef477d1 fixed a hang on exit by unconditionally unref'ing the
reconnect timer.  It worked but it's ultimately a cheap hack.

As of the previous commit, uhura explicitly tracks whether the socket
is ref'd or unref'd.  This commit extends that to the reconnect timer.

I'm afraid there's no regression test.  Reliably testing for event loop
stalls is pretty much impossible with mocha.