lib: fix SendEmitter#attach() queued cmd dispatch
Mea culpa, it should call Array#shift() instead of Array#unshift().

No regression test yet.  The only function that pushes actions onto
the queue is SendEmitter#_send() and then only when it's not attached
to a socket yet.  It turns out that's quite hard to capture reliably in
a test.