lib: remove maxQueueLength property
strong-agent and nodefly-{apiserver,collector} don't use it and it's
undocumented.

Its behavior of indiscriminately dropping the oldest messages is also
unlikely to be what you want because it destroys the order of events.

Last but not least, the section of code that it's part of is due to
change significantly in the near future and I don't want to expend
effort keeping something around that is unused.  Ergo, remove it.